// src/app/models/BaharNumber.model.ts
export class BaharNumber{
  // singleNumberId: number;
  // singleNumber: number;
  // numberCombinations?: {
  //   numberCombinationId: number,
  //   tripleNumber: number,
  //   visibleTripleNumber: string
  //   quantity?: number
  // }[];
  baharNumberId: number;
  baharNumber: number;
  // single_order: number;
}